Refactor: move Favorite data class to core domain
