+++ core/data/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
   id("bazzmovies.android.library")
-  id("kotlin-parcelize")
 }
 
 android.namespace = "com.waffiq.bazz_movies.core.data"
