+++ core/database/src/main/kotlin/com/waffiq/bazz_movies/core/database/domain/repository/IDatabaseRepository.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.core.database.domain.repository
 
-import com.waffiq.bazz_movies.core.database.data.model.Favorite
 import com.waffiq.bazz_movies.core.database.utils.DbResult
+import com.waffiq.bazz_movies.core.domain.Favorite
 import kotlinx.coroutines.flow.Flow
 
 interface IDatabaseRepository {
