+++ core/database/src/main/kotlin/com/waffiq/bazz_movies/core/database/domain/usecase/local_database/LocalDatabaseUseCase.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.core.database.domain.usecase.local_database
 
-import com.waffiq.bazz_movies.core.database.data.model.Favorite
 import com.waffiq.bazz_movies.core.database.utils.DbResult
+import com.waffiq.bazz_movies.core.domain.Favorite
 import kotlinx.coroutines.flow.Flow
 
 interface LocalDatabaseUseCase {
