+++ core/domain/src/main/kotlin/com/waffiq/bazz_movies/core/domain/FavoriteModel.kt
@@ -1,5 +1,8 @@
 package com.waffiq.bazz_movies.core.domain
 
+/**
+ * Used for TMDB POST method to favorite
+ */
 data class FavoriteModel(
   val mediaType: String,
   val mediaId: Int,
