+++ core/domain/src/main/kotlin/com/waffiq/bazz_movies/core/domain/Post.kt
@@ -1,5 +1,8 @@
 package com.waffiq.bazz_movies.core.domain
 
+/**
+ * Used as return value from general POST method
+ */
 data class Post(
   val success: Boolean? = null,
   val statusCode: Int? = null,
