+++ core/domain/src/main/kotlin/com/waffiq/bazz_movies/core/domain/WatchlistModel.kt
@@ -1,5 +1,8 @@
 package com.waffiq.bazz_movies.core.domain
 
+/**
+ * Used for TMDB POST method to watchlist
+ */
 data class WatchlistModel(
   val mediaType: String,
   val mediaId: Int,
